+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add searchable genes list.
 
 ### Changed
+- Lazy load HiGlass (and PIXI.js) from absolute URLs (unpkg) to avoid dynamic script import issues with relative paths after bundling.
 
 ## [1.1.0](https://www.npmjs.com/package/vitessce/v/1.1.0) - 2020-11-17